requirements.txt
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Опциональные зависимости (для дополнительных функций)
# Для веб-дашборда (если будет добавлен):
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Для базы данных (если будет добавлена):
# aiosqlite>=0.19.0
# sqlalchemy>=2.0.0

